// types/customer.ts
export interface Customer {
  id: number;

  firstName: string;
  lastName: string;
  gender: string;

  age: number;
  city: string;
  country: string;

  email: string;
  phone: string;

  height: number;
  income: number;

  college: string;
  degree: string;

  company: string;
  designation: string;

  maritalStatus: string;

  religion: string;
  caste: string;

  languages: string[];

  siblings: number;

  wantKids: string;
  openToRelocate: string;
  openToPets: string;

  
}